Renames and docstrings
