+++ pylintrc
@@ -11,7 +11,7 @@ max-locals=25
 good-names=ax,im,ix,a,b,c,h,_,n,p,s,fs,i,j,k,t,x,y,z,lc,rc,df, il, xl, tr, tr_, dx, dy, dz, q, x_, h_, hs, ds, h, v
 
 [MESSAGE CONTROL]
-disable=no-value-for-parameter,no-self-use,too-few-public-methods,unsubscriptable-object,no-method-argument,no-member,arguments-differ,len-as-condition,keyword-arg-before-vararg,too-many-locals,logging-format-interpolation, no-name-in-module
+disable=no-value-for-parameter,no-self-use,too-few-public-methods,unsubscriptable-object,no-method-argument,no-member,arguments-differ,len-as-condition,keyword-arg-before-vararg,too-many-locals,logging-format-interpolation, no-name-in-module, import-error
 
 [TYPECHECK]
 ignored-modules=numpy, numba
