--- seismiqb/src/_const.py
@@ -1,4 +1,0 @@
-""" Container for constant values used in SeismiQB. """
-
-FILL_VALUE = -999999 # placeholder value for horizon depth inside labels dictionary
-FILL_VALUE_MAP = -999999999 # placeholder value for horizon depth inside depth map
